test: ReservationTest 테스트 코드 리팩터링
- ReservationTest에 있던 ReservationTime 관련 테스트를 별도 단위 테스트로 분리
