+++ src/test/java/roomescape/ReservationTest.java
@@ -1,11 +1,11 @@
 package roomescape;
 
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
 import java.time.LocalDate;
 import java.time.LocalTime;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 import roomescape.model.Reservation;
+import roomescape.model.ReservationTime;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
@@ -18,7 +18,7 @@ class ReservationTest {
         String invalidName = "";
         long validId = 1;
         LocalDate validDate = LocalDate.now();
-        LocalTime validTime = LocalTime.now();
+        ReservationTime validTime = new ReservationTime(1L, LocalTime.now());
 
         // when & then
         assertThatThrownBy(() -> new Reservation(validId, invalidName, validDate, validTime))
@@ -32,24 +32,10 @@ class ReservationTest {
         LocalDate invalidDate = null;
         String validName = "포비";
         long validId = 1;
-        LocalTime validTime = LocalTime.now();
+        ReservationTime validTime = new ReservationTime(1L, LocalTime.now());
 
         // when & then
         assertThatThrownBy(() -> new Reservation(validId, validName, invalidDate, validTime))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    @DisplayName("시간 값이 null이면 예외가 발생한다")
-    void timeExceptionTest() {
-        // given
-        LocalTime invalidTime = null;
-        String validName = "포비";
-        long validId = 1;
-        LocalDate validDate = LocalDate.now();
-
-        // when & then
-        assertThatThrownBy(() -> new Reservation(validId, validName, validDate, invalidTime))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 }
